Specified Gem versions; added bin/setup script.
Having each Gem in the `Gemfile` with an explicit specification of which
versions are allowed and what minimum version must be installed will be
useful as the underlying Gems continue to undergo development. Doing so
also improves the repeatability of bundling the app; it is less likely
that bringing the app up on a new developer’s system will have odd
version conflicts and mismatches if the `Gemfile` is this explicit.

Adding a `bin/setup` file codifies our understanding of what is needed to
get the application into a known initial state. As dependencies requiring
setup are added to the app (non-toy database, Redis or similar caching,
etc), the commands to set those to an initial known-good state should
also be added to the bin/setup file.